+++ src/app/(components)/_base/dialog.tsx
@@ -54,7 +54,7 @@ export function BaseDialog({
         </HeadlessTransitionChild>
 
         <HeadlessTransitionChild
-          className="fixed inset-0 z-[50] w-screen overflow-y-auto pt-6 sm:pt-0"
+          className="fixed inset-0 z-[5] w-screen overflow-y-auto pt-6 sm:pt-0"
           enter="ease-out duration-100"
           enterFrom="opacity-0 translate-y-12 sm:translate-y-0"
           enterTo="opacity-100 translate-y-0"